+++ source/TrainManager/TrackFollowingObject/AI/TrackFollowingObject.AI.cs
@@ -77,11 +77,11 @@ namespace TrainManager.Trains
 			// Set the state quantity of the train.
 			if (DeltaPosition < 0)
 			{
-				Train.Handles.Reverser.Driver = ReverserPosition.Reverse;
+				Train.Handles.Reverser.Driver = (int)ReverserPosition.Reverse;
 			}
 			else if (DeltaPosition > 0)
 			{
-				Train.Handles.Reverser.Driver = ReverserPosition.Forwards;
+				Train.Handles.Reverser.Driver = (int)ReverserPosition.Forwards;
 			}
 
 			Train.Handles.Reverser.Actual = Train.Handles.Reverser.Driver;
